split into two, menu not implemented yet
